so dcr and application completed

+++ lib/screen/student/daily_class_report/widgets/attendence_bar.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class AttendanceBar extends StatelessWidget {
-  final int percentage;
+  final double percentage;
 
   const AttendanceBar({super.key, required this.percentage});
